@@ -9,7 +9,7 @@ int main(int argc, char *argv[])
 {
         QApplication a(argc, argv);
         if(check_uid()){
-                dialog(0,"You must to have root privileage").exec();
+                Dialog("", "You must to have root privileage").exec();
                 exit(1);
         }
         mainwindow w;
